Finish hashset

+++ hashmap.js
@@ -43,7 +43,6 @@ class HashMap {
         const hashCode = this.hash(key);
         let bucket = this.map[hashCode];
         let point = (bucket) ? bucket.head: null;
-        // console.log(bucket);
         while (point) {
             if (point.data[0] === key) {
                 return point.data[1];
@@ -67,12 +66,17 @@ class HashMap {
     remove(key) {
         const hashCode = this.hash(key);
         let bucket = this.map[hashCode];
-        let point = bucket.head;
+        let point = (bucket) ? bucket.head: null;
         let index = 0;
         const pair = [key, this.get(key)];
         while (point) {
             if (pair[0] === point.data[0] && pair[1] === point.data[1]) {
-                bucket.removeAt(index);
+                if (bucket.head[0] === bucket.tail[0]) {
+                    this.map[hashCode] = undefined;
+                }
+                else {
+                    bucket.removeAt(index);
+                }
                 return true;
             }
             index += 1;
@@ -139,7 +143,6 @@ class HashMap {
         let key;
         let value;
         let newArrayBucket;
-        let newArrayPoint;
         if (this.length >= this.capacity * this.loadFactor) {
             this.capacity *= 2;
             for (let i=0; i<this.map.length; i++) {
@@ -159,9 +162,130 @@ class HashMap {
                     point = point.next;
                 }
             }
+            this.map = newArray;
         }
-        this.map = newArray;
     }
 }
 
-export default HashMap;
+class HashSet {
+    constructor(capacity=16) {
+        this.capacity = capacity;
+        this.array = [];
+        this.loadFactor = 0.75;
+        this.length = 0;
+    }
+    hash(key) {
+        let hashCode = 0;
+        
+        const primeNumber = 31;
+        for (let i = 0; i < key.length; i++) {
+            hashCode = primeNumber * hashCode + key.charCodeAt(i);
+        }
+     
+        return hashCode % this.capacity;
+    }
+    set(key) {
+        const hashCode = this.hash(key);
+        // check if bucket exists before making a bucket
+        if (!this.array[hashCode]) {
+            this.array[hashCode] = new LinkedList(new Node(key));
+        }
+        // find the key in the bucket to overwrite or add if unique key
+        else {
+            let bucket = this.array[hashCode];
+            let point = bucket.head;
+            while (point) {
+                if (point === key) {
+                    return;
+                }
+                point = point.next;
+            }
+            bucket.append(key);
+        }
+        this.grow();
+        this.length++;
+    }
+    has(key) {
+        const hashCode = this.hash(key);
+        let bucket = this.array[hashCode];
+        let point = (bucket) ? bucket.head : null;
+        while (point) {
+            if (point.data === key) {
+                return true;
+            }
+            point = point.next;
+        }
+        return false;
+    }
+    remove(key) {
+        const hashCode = this.hash(key);
+        let bucket = this.array[hashCode];
+        let point = bucket.head;
+        let index = 0;
+        while (point) {
+            if (point.data === key) {
+                if (bucket.head === bucket.tail) {
+                    this.array[hashCode] = undefined;
+                }
+                else {
+                    bucket.removeAt(index);
+                }
+                return true;
+            }
+            index += 1;
+            point = point.next;
+        }
+        return false;
+    }
+    clear() {
+        this.array = [];
+        this.length = 0;
+        this.capacity = 16;
+    }
+    keys() {
+        let keyArray = [];
+        let bucket;
+        let point;
+        let key;
+        for (let i=0; i<this.array.length; i++) {
+            bucket = this.array[i];
+            point = (bucket) ? bucket.head: null;
+            while (point) {
+                key = point.data;
+                keyArray.push(key);
+                point = point.next;
+            }
+        }
+        return keyArray;
+    }
+    grow() {
+        let newArray = [];
+        let bucket;
+        let point;
+        let hashCode;
+        let key;
+        let newArrayBucket;
+        if (this.length >= this.capacity * this.loadFactor) {
+            this.capacity *= 2;
+            for (let i=0; i<this.array.length; i++) {
+                bucket = this.array[i];
+                point = (bucket) ? bucket.head: null;
+                while (point) {
+                    key = point.data;
+                    hashCode = this.hash(key);
+                    if (!newArray[hashCode]) {
+                        newArray[hashCode] = new LinkedList(new Node(key));
+                    }
+                    else {
+                        newArrayBucket = newArray[hashCode];
+                        newArrayBucket.append(key);
+                    }
+                    point = point.next;
+                }
+            }
+            this.array = newArray;
+        }
+    }
+}
+
+export {HashMap, HashSet};
